requirements.txt
Django~=2.2.4
djangorestframework~=3.12.4
pip~=21.2.4
wheel~=0.36.2
setuptools~=57.0.0